pyproject.toml
[project]
name = "python-workshop-am"
version = "0.1.0"
description = "A simple Link Tree-style showcase app for professors and academics"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "pyyaml>=6.0",
    "gunicorn>=21.0.0"
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
